base64: replace old implementation with new one
