--- bindings/new_base64.lua
@@ -1,123 +1,0 @@
--- Base64-encoding
--- Sourced from http://en.wikipedia.org/wiki/Base64
---
--- Copyright (c) 2012, Daniel Lindsley
--- All rights reserved.
---
--- Modified by: Max Bruckner (FSMaxB)
--- * remove unnecessary "require"
--- * remove __ attributes
--- * rename "to_base64" -> "encode" and "from_base64" -> "decode"
--- * make it a loadable module via require
--- * fix use of the modulo operator (math.mod -> %)
---
--- Redistribution and use in source and binary forms, with or without
--- modification, are permitted provided that the following conditions are met:
---
--- * Redistributions of source code must retain the above copyright notice, this
---   list of conditions and the following disclaimer.
--- * Redistributions in binary form must reproduce the above copyright notice,
---   this list of conditions and the following disclaimer in the documentation
---   and/or other materials provided with the distribution.
--- * Neither the name of the base64 nor the names of its contributors may be
---   used to endorse or promote products derived from this software without
---   specific prior written permission.
---
--- THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
--- ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
--- WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
--- DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
--- FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
--- DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
--- SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
--- CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
--- OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
--- OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-
-local base64 = {}
-
-local index_table = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/'
-
-
-local function to_binary(integer)
-    local remaining = tonumber(integer)
-    local bin_bits = ''
-
-    for i = 7, 0, -1 do
-        local current_power = math.pow(2, i)
-
-        if remaining >= current_power then
-            bin_bits = bin_bits .. '1'
-            remaining = remaining - current_power
-        else
-            bin_bits = bin_bits .. '0'
-        end
-    end
-
-    return bin_bits
-end
-
-local function from_binary(bin_bits)
-    return tonumber(bin_bits, 2)
-end
-
-
-function base64.encode(to_encode)
-    local bit_pattern = ''
-    local encoded = ''
-    local trailing = ''
-
-    for i = 1, string.len(to_encode) do
-        bit_pattern = bit_pattern .. to_binary(string.byte(string.sub(to_encode, i, i)))
-    end
-
-    -- Check the number of bytes. If it's not evenly divisible by three,
-    -- zero-pad the ending & append on the correct number of ``=``s.
-    if (string.len(bit_pattern) % 3) == 2 then
-        trailing = '=='
-        bit_pattern = bit_pattern .. '0000000000000000'
-    elseif (string.len(bit_pattern) % 3) == 1 then
-        trailing = '='
-        bit_pattern = bit_pattern .. '00000000'
-    end
-
-    for i = 1, string.len(bit_pattern), 6 do
-        local byte = string.sub(bit_pattern, i, i+5)
-        local offset = tonumber(from_binary(byte))
-        encoded = encoded .. string.sub(index_table, offset+1, offset+1)
-    end
-
-    return string.sub(encoded, 1, -1 - string.len(trailing)) .. trailing
-end
-
-
-function base64.decode(to_decode)
-    local padded = to_decode:gsub("%s", "")
-    local unpadded = padded:gsub("=", "")
-    local bit_pattern = ''
-    local decoded = ''
-
-    for i = 1, string.len(unpadded) do
-        local char = string.sub(to_decode, i, i)
-        local offset, _ = string.find(index_table, char)
-        if offset == nil then
-             error("Invalid character '" .. char .. "' found.")
-        end
-
-        bit_pattern = bit_pattern .. string.sub(to_binary(offset-1), 3)
-    end
-
-    for i = 1, string.len(bit_pattern), 8 do
-        local byte = string.sub(bit_pattern, i, i+7)
-        decoded = decoded .. string.char(from_binary(byte))
-    end
-
-    local padding_length = padded:len()-unpadded:len()
-
-    if (padding_length == 1 or padding_length == 2) then
-        decoded = decoded:sub(1,-2)
-    end
-    return decoded
-end
-
-return base64
